refactor: rename createElemDom.ts to utils.ts, rename folder to utils, move folder to project root
